+++ src/app/page.tsx
@@ -12,7 +12,7 @@ import Footer from '@/components/Footer';
 import Header from '@/components/Header';
 import HeroSection from '@/components/HeroSection';
 import LoginModal from '@/components/LoginModal';
-import { tokenManager } from '@/lib/auth/tokenManager';
+import { tokenManager } from '@/lib/api/tokenManager';
 
 import styles from './page.module.css';
